requirements.txt
aiohttp
discum
slack_sdk
gspread==5.7.2
oauth2client==4.1.3
python-dotenv==0.21.1
requests==2.28.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
requests-toolbelt==0.9.1
Flask
aiohttp